Added damage/health system
Created a script called EnemyDestructableHealth that holds a health
amount. Attach it to any enemy or destructable object and tag the object
as enemy or destructable and the fireball will do damage on collision.
Then in the objects controller script, see if the healt is depleted and
if it is, destroy.